Add and test support for views in packet cache.
The packet cache data buckets are now set up in an unordered map
addressed by the view name.

Doing this also makes sure that, if the network configuration of a view
changes, the cache contents are still valid as long as there is no
change in the zones found in that view.