Split retrieval of view from retrieval of variant.
First, getting the view from the netmask can only be performed once,
rather than every iteration in the chopoff loop until the zone is found.

Second, this will allow code which only cares about the view (in a
non-zone name context) to be able to retrieve it easily.

This also hints that the Netmask->View retrieval might be put in a
different place than the zone cache eventually.