Redo the interaction between views and packet cache.
Rather than attempting to tie cache data to the networks addressing the
views, tie them to the view name directly. The packet cache data buckets
are now set up in an unordered map addressed by the view name.

Doing this eliminates the need to check for matching network masks, and
also makes sure that, if the network configuration of a view changes,
the cache contents are still valid as long as there is no change in the
zones found in that view.