Rename some of the views-related pdnsutil commands.
Most pdnsutil commands are written as "VERB-NOUN" (or "ACTION-OBJECT" if
you prefer), and the networks and view listing commands had been
initially written the other way ("NOUN-VERB").

The commands with one verb and two objects (view-{add,del}-zone) are
left unchanged because this is the only non-awkward way to spell them;
this also makes a point in favour of rewriting the commands with the
object type they apply to first, but that's out of the scope of this PR,
which has already exhausted its bikeshed quota.